requirements.txt
matplotlib==3.10.7
numpy==2.3.4
Pillow==12.0.0
sionna==1.2.1
sionna_rt==1.2.1
tensorflow==2.14.0
torch==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.4
